+++ test/input_schema.test.ts
@@ -14,7 +14,7 @@ describe('input_schema.json', () => {
                 properties: {
                     myField: {
                         title: 'Field title',
-                        type: ['object', 'array', 'string', 'integer', 'boolean'],
+                        type: ['object', 'array', 'string', 'integer', 'number', 'boolean'],
                         description: 'Some description ...',
                         editor: 'json',
                     },
@@ -39,6 +39,12 @@ describe('input_schema.json', () => {
                         description: 'Some description ...',
                         editor: 'fileupload',
                     },
+                    myfield5: {
+                        title: 'Array fileupload title',
+                        type: 'array',
+                        description: 'Some description...',
+                        editor: 'fileupload',
+                    },
                 },
             };
 
@@ -54,7 +60,7 @@ describe('input_schema.json', () => {
                 properties: {
                     myField: {
                         title: 'Field title',
-                        type: ['object', 'array', 'string', 'integer', 'boolean'],
+                        type: ['object', 'array', 'string', 'integer', 'number', 'boolean'],
                         description: 'Some description ...',
                         editor: 'json',
                     },
@@ -72,7 +78,7 @@ describe('input_schema.json', () => {
                 properties: {
                     myField: {
                         title: 'Field title',
-                        type: ['object', 'array', 'string', 'integer', 'boolean'],
+                        type: ['object', 'array', 'string', 'integer', 'number', 'boolean'],
                         description: 'Some description ...',
                         editor: 'json',
                     },
@@ -92,7 +98,7 @@ describe('input_schema.json', () => {
                 properties: {
                     myField: {
                         title: 'Field title',
-                        type: ['object', 'array', 'string', 'integer', 'boolean'],
+                        type: ['object', 'array', 'string', 'integer', 'number', 'boolean'],
                         description: 'Some description ...',
                         editor: 'textfield',
                     },
@@ -841,6 +847,7 @@ describe('input_schema.json', () => {
                 const types = [
                     { type: 'string', editor: 'textfield' },
                     { type: 'integer', editor: 'number' },
+                    { type: 'number', editor: 'number' },
                     { type: 'boolean', editor: 'checkbox' },
                     { type: 'array', editor: 'json' },
                     { type: 'object', editor: 'json' },
@@ -869,6 +876,7 @@ describe('input_schema.json', () => {
                 const types = [
                     { type: 'string', editor: 'textfield' },
                     { type: 'integer', editor: 'number' },
+                    { type: 'number', editor: 'number' },
                     { type: 'boolean', editor: 'checkbox' },
                     { type: 'array', editor: 'json' },
                     { type: 'object', editor: 'json' },
@@ -912,6 +920,151 @@ describe('input_schema.json', () => {
                         `Input schema is not valid (Field schema.properties.myField.default must be ${type.type})`,
                     );
                 });
+            });
+        });
+
+        describe('validate pattern regexps', () => {
+            it('should accept valid regexp', () => {
+                const schema = {
+                    title: 'Test input schema',
+                    type: 'object',
+                    schemaVersion: 1,
+                    properties: {
+                        myField: {
+                            title: 'Field title',
+                            type: 'string',
+                            description: 'Some description ...',
+                            editor: 'textfield',
+                            pattern: '^[A-Z]{3}$',
+                        },
+                        objectField: {
+                            title: 'Object field',
+                            type: 'object',
+                            description: 'Some description ...',
+                            editor: 'json',
+                            patternKey: '^[a-z]+$',
+                            patternValue: '^[0-9]+$',
+                        },
+                        arrayField: {
+                            title: 'Array field',
+                            type: 'array',
+                            description: 'Some description ...',
+                            editor: 'json',
+                            patternKey: '^[a-z]+$',
+                            patternValue: '^[0-9]+$',
+                        },
+                    },
+                };
+
+                expect(() => validateInputSchema(validator, schema)).not.toThrow();
+            });
+
+            it('should throw error on invalid pattern regexp', () => {
+                const schema = {
+                    title: 'Test input schema',
+                    type: 'object',
+                    schemaVersion: 1,
+                    properties: {
+                        myField: {
+                            title: 'Field title',
+                            type: 'string',
+                            description: 'Some description ...',
+                            editor: 'textfield',
+                            pattern: '[A-Z{3}', // invalid regexp
+                        },
+                    },
+                };
+
+                expect(() => validateInputSchema(validator, schema)).toThrow(
+                    'Input schema is not valid (The regular expression "[A-Z{3}" in field schema.properties.myField.pattern must be valid.)',
+                );
+            });
+
+            it('should throw error on invalid patternKey regexp', () => {
+                const schema = {
+                    title: 'Test input schema',
+                    type: 'object',
+                    schemaVersion: 1,
+                    properties: {
+                        objectField: {
+                            title: 'Object field',
+                            type: 'object',
+                            description: 'Some description ...',
+                            editor: 'json',
+                            patternKey: '[a-z+$', // invalid regexp
+                            patternValue: '^[0-9]+$',
+                        },
+                    },
+                };
+
+                expect(() => validateInputSchema(validator, schema)).toThrow(
+                    'Input schema is not valid (The regular expression "[a-z+$" in field schema.properties.objectField.patternKey must be valid.)',
+                );
+            });
+
+            it('should throw error on invalid patternValue regexp', () => {
+                const schema = {
+                    title: 'Test input schema',
+                    type: 'object',
+                    schemaVersion: 1,
+                    properties: {
+                        objectField: {
+                            title: 'Object field',
+                            type: 'object',
+                            description: 'Some description ...',
+                            editor: 'json',
+                            patternKey: '^[a-z]+$',
+                            patternValue: '^[0-9+$', // invalid regexp
+                        },
+                    },
+                };
+
+                expect(() => validateInputSchema(validator, schema)).toThrow(
+                    'Input schema is not valid (The regular expression "^[0-9+$" in field schema.properties.objectField.patternValue must be valid.)',
+                );
+            });
+
+            it('should throw error on not safe regexp', () => {
+                const invalidRegexps = [
+                    '(a+)+$',
+                    '^(a|a?)+$',
+                    '^(a|a*)+$',
+                    '^(a|a+)+$',
+                    '^(a?)+$',
+                    '^(a*)+$',
+                    '^(a+)*$',
+                    '^(a|aa?)+$',
+                    '^(a|aa*)+$',
+                    '^(a|a+)*$',
+                    '^(a|a?)*$',
+                    '^(a|a*)*$',
+                    '^(a?)*$',
+                    '^(a*)*$',
+                    '^(a+)?$',
+                    '^(a*)?$',
+                    'a*b*c*d*e*f*g*h*i*j*k*l*m*n*o*p*q*r*s*t*u*v*w*x*y*z*',
+                ];
+
+                for (const pattern of invalidRegexps) {
+                    const schema = {
+                        title: 'Test input schema',
+                        type: 'object',
+                        schemaVersion: 1,
+                        properties: {
+                            myField: {
+                                title: 'Field title',
+                                type: 'string',
+                                description: 'Some description ...',
+                                editor: 'textfield',
+                                pattern,
+                            },
+                        },
+                    };
+
+                    expect(() => validateInputSchema(validator, schema)).toThrow(
+                        `Input schema is not valid (The regular expression "${pattern}" in field schema.properties.myField.pattern may cause excessive backtracking or be unsafe to execute.)`,
+                    );
+                }
             });
         });
     });